+++ src/Control/Effect/Choose.hs
@@ -13,14 +13,14 @@ module Control.Effect.Choose
 , Has
 ) where
 
-import Control.Carrier
+import {-# SOURCE #-} Control.Carrier
 import Control.Effect.Empty
 import Data.Bool (bool)
 import Data.List.NonEmpty (NonEmpty (..))
 import qualified Data.Semigroup as S
 import GHC.Generics (Generic1)
 
-data Choose m k
+newtype Choose m k
   = Choose (Bool -> m k)
   deriving (Functor, Generic1)