+++ src/Control/Effect/Empty.hs
@@ -8,7 +8,7 @@ module Control.Effect.Empty
 , Has
 ) where
 
-import Control.Carrier
+import {-# SOURCE #-} Control.Carrier
 import GHC.Generics (Generic1)
 
 -- | An effect modelling nondeterminism without choice.